All in redred file
I addad all info to the rrscan class and deleted rrMOKE class. rrScans class now also in redred file. RRMOKE now just som test skript.
